refactor(pm): split service/script.rs into exec / node_gyp / lifecycle
The 983-line file mixed three responsibilities; it is now a script/
directory module: exec.rs (command construction + execution
primitives), node_gyp.rs (toolchain bootstrap), lifecycle.rs
(run_lifecycle/run_script/run_in_layers orchestration and the sink
types). Pure move — mod.rs re-exports the public surface so no caller
changed; tests moved with the code they exercise.
